finetune/requirements.txt
transformers
datasets
evaluate
scikit-learn
pandas
numpy
pyyaml
tqdm
accelerate
torch
